IGNITE-17170 Drop deprecated package
