+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/commands/node/NodeUrlMixin.java
@@ -28,7 +28,7 @@ import jakarta.inject.Inject;
 import java.net.URL;
 import org.apache.ignite.internal.cli.NodeNameRegistry;
 import org.apache.ignite.internal.cli.core.converters.UrlConverter;
-import org.apache.ignite.internal.cli.deprecated.IgniteCliException;
+import org.apache.ignite.internal.cli.core.exception.IgniteCliException;
 import org.jetbrains.annotations.Nullable;
 import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Option;
