+++ modules/cli/src/main/java/org/apache/ignite/internal/cli/core/exception/IgniteCliException.java
@@ -15,19 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cli.deprecated;
-
-import org.junit.jupiter.api.BeforeAll;
+package org.apache.ignite.internal.cli.core.exception;
 
 /**
- * Base class for any CLI tests.
+ * Top level runtime exception for throwing the error message to user.
  */
-public abstract class AbstractCliTest {
+public class IgniteCliException extends RuntimeException {
+    /** Serial version uid. */
+    private static final long serialVersionUID = 0L;
+
     /**
-     * Sets up a dumb terminal before tests.
+     * Creates a new instance of {@code IgniteCliException} with the given {@code msg}.
+     *
+     * @param msg Detailed message.
      */
-    @BeforeAll
-    static void beforeAll() {
-        System.setProperty("org.jline.terminal.dumb", "true");
+    public IgniteCliException(String msg) {
+        super(msg);
+    }
+
+    /**
+     * Creates a new instance of {@code IgniteCliException} with the given {@code msg} and {@code cause}.
+     *
+     * @param msg   Detailed message.
+     * @param cause Cause.
+     */
+    public IgniteCliException(String msg, Throwable cause) {
+        super(msg, cause);
     }
 }
